module_2_defect/requirements.txt
pygltflib==1.16.5
pypdf>=4.1.0
Pillow>=10.0.0
Flask-SQLAlchemy>=3.0.0
psycopg2-binary>=2.9.0
